read controller parameter "default" from yaml file
When reading controller configurations, "default" is ignored as
opposed to what is currently written in the documentation.
MoveItFakeControllerManager now maintains map of controller name and
ControllerState where "default" is initialized as "false" or whatever
is given in the configuration.
MoveItFakeControllerManager::getControllerState() now returns the
state that is held in that map instead of a default one.